Connection Joint Update
This fixes some of the warping issues prevalent on the old iterations of the connection joints. Personally it made these  much more printable however little testing has been done. It is really just a fillet added to the exterior except for the printed corners.

File Structure has been reworked as well.