Major Y Axis Linear Rail Mount Rework
The old system was built on using MGN 15 rails, which is completely overkill for the task at hand especially when mounted to alluminium beams. therefore this has been replaced with connection joints designed to work with MGN 12 rails instead. This required an entire rework of the y axis assembly, however with that out of the way I am happy with how it works together. The y axis assembly require 520mm mgn 12 rails x 2, preferably I would like to get this size down to 500mm for example so people can buy the rails no cutting required, however that also means a 20mm reduction in cut volume.