Y-Axis Idler Reimproved
The previous designed roller likes to bend a lot and is ineffective, and V2 carries a lot of the problems that V1 had, but does solve the bending. V3 is untested but should be a good compromise of both, can be printed without supports, but is quite a heavy model.

Quite a bit of inactivity as my workspace is being upgraded, and parts need to be bought, primarily printer belts.

The Z axis motor mounts for left and right have been merged into one file, and there has been hole finder for the Y Axis motor mounts. No update to the fusion file as for now, that can come within the coming weeks.