Toolhead addition + X Axis Idler Tensioner
This update adds the toolhead which allows the laser to be mounted. The laser type chosen was a LASER TREE 10W Laser Module with Air Assist, which has direct built-in air assist great for the operation.

The X Axis Idler Joint has also been completely reworked to add the same tensioner system as is installed on the Y axis, this does mean we lose about 10mm of build volume but that can be readded later on by improvements in the design.