Z Axis Motor Mount Changes
This is a pretty wasteful way to find out where to punch and tap the Z axis motor mount holes, but it does work, and does not require the use of a tape measure
or much thinking, meaning it is a reliable way for novice people to find the center holes. If you are more advanced measuring out the center point and mounting
based on that would be a more cost-saving measure, but that is on you.

CHANGELOG
- Designed Z Motor Mounts
- Designed Z Axis Top
- Revised Toolhead