Motor Mount and Idler Pulley setup Revision
Now that the Y to X design has been mostly finalized and has been confirmed to work, I have now redesigned them to increase strength to combat issues I was having with the connectors snapping when disassembling the axis. This now includes a loft in both the idler pulley, and the X axis idler motor. There is quite a bit of wasted build space with the idler pulley so that is WIP, but the motor mount should be good to go.